move sticky tab page context to separate file
