+++ src/lib/Components/StickyTabPage/StickyTabPage.tsx
@@ -1,7 +1,7 @@
 import { LegacyNativeModules } from "lib/NativeModules/LegacyNativeModules"
 import { unsafe__getSelectedTab } from "lib/store/GlobalStore"
 import { Schema } from "lib/utils/track"
-import { GlobalState, useGlobalState } from "lib/utils/useGlobalState"
+import { useGlobalState } from "lib/utils/useGlobalState"
 import { useScreenDimensions } from "lib/utils/useScreenDimensions"
 import { color } from "palette"
 import React, { useMemo, useRef, useState } from "react"
@@ -9,35 +9,10 @@ import { View } from "react-native"
 import Animated from "react-native-reanimated"
 import { useTracking } from "react-tracking"
 import { useAnimatedValue } from "./reanimatedHelpers"
+import { StickyTabPageContext } from "./SitckyTabPageContext"
 import { SnappyHorizontalRail } from "./SnappyHorizontalRail"
 import { StickyTabPageFlatListContext } from "./StickyTabPageFlatList"
 import { StickyTabPageTabBar } from "./StickyTabPageTabBar"
-
-const StickyTabPageContext = React.createContext<{
-  staticHeaderHeight: Animated.Node<number> | null
-  stickyHeaderHeight: Animated.Node<number> | null
-  headerOffsetY: Animated.Value<number>
-  tabLabels: string[]
-  activeTabIndex: GlobalState<number>
-  setActiveTabIndex(index: number): void
-}>(
-  __TEST__
-    ? {
-        staticHeaderHeight: new Animated.Value(0),
-        stickyHeaderHeight: new Animated.Value(0),
-        headerOffsetY: new Animated.Value(0),
-        tabLabels: ["test"],
-        // tslint:disable-next-line:no-empty
-        activeTabIndex: { current: 0, set() {}, useUpdates() {} },
-        // tslint:disable-next-line:no-empty
-        setActiveTabIndex() {},
-      }
-    : (null as any)
-)
-
-export function useStickyTabPageContext() {
-  return React.useContext(StickyTabPageContext)
-}
 
 interface TabProps {
   initial?: boolean
